+++ django/applications/catmaid/control/node.py
@@ -5,6 +5,7 @@ from collections import defaultdict
 from datetime import datetime
 import itertools as itertools
 
+from django.conf import settings
 from django.db import connection
 from django.http import HttpResponse
 from django.contrib.auth.models import User
@@ -49,7 +50,8 @@ def node_list_tuples(request, project_id=None, provider=None):
     atnid = int(request.POST.get('atnid', -1))
     for p in ('top', 'left', 'bottom', 'right', 'z1', 'z2'):
         params[p] = float(request.POST.get(p, 0))
-    params['limit'] = 5000  # Limit the number of retrieved treenodes within the section
+    # Limit the number of retrieved treenodes within the section
+    params['limit'] = settings.NODE_LIST_MAXIMUM_COUNT
     params['project_id'] = project_id
     includeLabels = (request.POST.get('labels', None) == 'true')